coral-hf-spaces (1)/coral-deploy/space1-detection/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
ultralytics
torch>=2.0.0